Add files via upload
Starting a new repository on my main github account. I'm tired of maintaining a work and a personal.

If anyone is here wondering about the history of this project, it was a collab between me and a coworker, Jonathon Gossett. But after the initial coding, he split to work for a different team, and I finished up the project and have made all changes since.